+++ src/main/java/com/braintreegateway/CreditCardVerification.java
@@ -35,6 +35,7 @@ public class CreditCardVerification {
     private Calendar createdAt;
     private RiskData riskData;
     private ThreeDSecureInfo threeDSecureInfo;
+    private String networkTransactionId;
 
     public CreditCardVerification(NodeWrapper node) {
         this.amount = node.findBigDecimal("amount");
@@ -77,6 +78,8 @@ public class CreditCardVerification {
         }
 
         this.createdAt = node.findDateTime("created-at");
+
+        this.networkTransactionId = node.findString("network-transaction-id");
     }
 
     public BigDecimal getAmount() {
@@ -166,5 +169,9 @@ public class CreditCardVerification {
 
     public Status getStatus() {
         return status;
+    }
+
+    public String getNetworkTransactionId() {
+        return networkTransactionId;
     }
 }